+++ CHANGELOG.md
@@ -7,10 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+## [0.3.0]
+
 ### Added
 
 - Create and update operation using bloom filters.
 - Text match querying using bloom filters.
+- Rake task to generate precision and recall statistics for text match querying.
 
 ### Fixed